refactor(events): extract event form
The new event screen now renders a shared form component instead of owning the full flow.

This keeps the route thin and moves the create form logic into one module.
